moved helpers out of commands.ts and added next steps to summary
