+++ .github/workflows/src/spec-gen-sdk-status.js
@@ -208,6 +208,13 @@ async function processResult({ checkRuns, core }) {
       ? "✅ All required spec-gen-sdk checks passed successfully!"
       : `❌ spec-gen-sdk checks failed for: ${specGenSdkFailedRequiredLanguages}`;
 
+  // Add next steps
+  if (state === CommitStatusState.FAILURE) {
+    summaryContent +=
+      "\n### Next Steps\n\n" +
+      `Please fix any issues in the the spec-gen-sdk checks for languages: ${specGenSdkFailedRequiredLanguages}.`;
+  }
+
   // Write to the summary page
   await writeToActionsSummary(summaryContent, core);
